Move parse_slice to lib.rs and remove comments from test file
